+++ src/constructors/test/styled.test.js
@@ -1,12 +1,11 @@
 // @flow
-import expect from 'expect'
 import styled from '../../index'
 import domElements from '../../utils/domElements'
 
 describe('styled', () => {
   it('should have all valid HTML5 elements defined as properties', () => {
     Object.keys(domElements).forEach(tagName => {
-      expect(styled[tagName]).toExist()
+      expect(styled[tagName]).toBeTruthy()
     })
   })
 })