+++ src/models/AbstractStyledComponent.js
@@ -1,7 +1,9 @@
 // @flow
-import { Component, PropTypes } from 'react'
+import { Component } from 'react'
+import PropTypes from 'prop-types'
 
 import { CHANNEL } from './ThemeProvider'
+import StyleSheet, { CONTEXT_KEY } from './StyleSheet'
 
 export default class AbstractStyledComponent extends Component {
   static isPrototypeOf: Function
@@ -15,4 +17,5 @@ export default class AbstractStyledComponent extends Component {
 
 AbstractStyledComponent.contextTypes = {
   [CHANNEL]: PropTypes.func,
+  [CONTEXT_KEY]: PropTypes.instanceOf(StyleSheet),
 }