+++ src/models/test/ThemeProvider.test.js
@@ -1,7 +1,7 @@
 // @flow
 /* eslint-disable react/no-multi-comp */
 import React from 'react'
-import expect from 'expect'
+import PropTypes from 'prop-types'
 import { shallow, render } from 'enzyme'
 import ThemeProvider, { CHANNEL } from '../ThemeProvider'
 
@@ -39,7 +39,7 @@ describe('ThemeProvider', () => {
       render() { return null }
     }
     Child.contextTypes = {
-      [CHANNEL]: React.PropTypes.object,
+      [CHANNEL]: PropTypes.object,
     }
 
     render(
@@ -66,7 +66,7 @@ describe('ThemeProvider', () => {
       render() { return null }
     }
     Child.contextTypes = {
-      [CHANNEL]: React.PropTypes.object,
+      [CHANNEL]: PropTypes.object,
     }
 
     render(
@@ -101,7 +101,7 @@ describe('ThemeProvider', () => {
       render() { return null }
     }
     Child.contextTypes = {
-      [CHANNEL]: React.PropTypes.object,
+      [CHANNEL]: PropTypes.object,
     }
 
     render(