+++ src/test/basic.test.js
@@ -1,10 +1,7 @@
 // @flow
 import React, { Component } from 'react'
-import expect from 'expect'
 import { shallow, mount } from 'enzyme'
-import jsdom from 'mocha-jsdom'
 
-import styleSheet from '../models/StyleSheet'
 import { resetStyled, expectCSSMatches } from './utils'
 
 let styled
@@ -21,18 +18,12 @@ describe('basic', () => {
     styled.div``
   })
 
-  it('should inject a stylesheet when a component is created', () => {
-    const Comp = styled.div``
-    shallow(<Comp />)
-    expect(styleSheet.injected).toBe(true)
-  })
-
-  it('should generate only component class by default', () => {
+  it('should not inject anything by default', () => {
     styled.div``
-    expectCSSMatches('.sc-a {}')
+    expectCSSMatches('')
   })
 
-  it('should generate only component class even if rendered if no styles are passed', () => {
+  it('should inject component class when rendered even if no styles are passed', () => {
     const Comp = styled.div``
     shallow(<Comp />)
     expectCSSMatches('.sc-a {}')
@@ -56,32 +47,98 @@ describe('basic', () => {
   })
 
   describe('jsdom tests', () => {
-    jsdom()
+    it('should pass the ref to the component', () => {
+      const Comp = styled.div``
 
-    let Comp
-    let WrapperComp
-    let wrapper
-
-    beforeEach(() => {
-      Comp = styled.div``
-      WrapperComp = class extends Component {
+      class Wrapper extends Component {
         testRef: any;
+        innerRef = (comp) => { this.testRef = comp }
+
         render() {
-          return <Comp innerRef={(comp) => { this.testRef = comp }} />
+          return <Comp innerRef={this.innerRef} />
         }
       }
 
-      wrapper = mount(<WrapperComp />)
+      const wrapper = mount(<Wrapper />)
+      const component = wrapper.find(Comp).first()
+
+      // $FlowFixMe
+      expect(wrapper.node.testRef).toBe(component.getDOMNode())
+      expect(component.find('div').prop('innerRef')).toBeFalsy()
     })
 
-    it('should pass ref to the component', () => {
+    class InnerComponent extends Component {
+      render() {
+        return null
+      }
+    }
+
+    it('should not leak the innerRef prop to the wrapped child', () => {
+      const OuterComponent = styled(InnerComponent)``
+
+      class Wrapper extends Component {
+        testRef: any;
+
+        render() {
+          return <OuterComponent innerRef={(comp) => { this.testRef = comp }} />
+        }
+      }
+
+      const wrapper = mount(<Wrapper />)
+      const innerComponent = wrapper.find(InnerComponent).first()
+
       // $FlowFixMe
-      expect(wrapper.node.testRef).toExist()
+      expect(wrapper.node.testRef).toBe(innerComponent.node)
+      expect(innerComponent.prop('innerRef')).toBeFalsy()
     })
 
-    it('should not pass innerRef to the component', () => {
+    it('should pass the full className to the wrapped child', () => {
+      const OuterComponent = styled(InnerComponent)``
+
+      class Wrapper extends Component {
+        render() {
+          return <OuterComponent className="test"/>
+        }
+      }
+
+      const wrapper = mount(<Wrapper />)
+      expect(wrapper.find(InnerComponent).prop('className'))
+        .toBe('test sc-a b')
+    })
+
+    it('should pass the innerRef to the wrapped styled component', () => {
+      const InnerComponent = styled.div``
+      const OuterComponent = styled(InnerComponent)``
+
+      class Wrapper extends Component {
+        testRef: any;
+        innerRef = (comp) => { this.testRef = comp }
+
+        render() {
+          return <OuterComponent innerRef={this.innerRef} />
+        }
+      }
+
+      const wrapper = mount(<Wrapper />)
+      const innerComponent = wrapper.find(InnerComponent).first()
+      const outerComponent = wrapper.find(OuterComponent).first()
+
       // $FlowFixMe
-      expect(wrapper.node.ref).toNotExist()
+      expect(wrapper.node.testRef).toBe(innerComponent.getDOMNode())
+
+      // $FlowFixMe
+      expect(innerComponent.prop('innerRef')).toBe(wrapper.node.innerRef)
+    })
+
+    it('should respect the order of StyledComponent creation for CSS ordering', () => {
+      const FirstComponent = styled.div`color: red;`
+      const SecondComponent = styled.div`color: blue;`
+
+      // NOTE: We're mounting second before first and check if we're breaking their order
+      shallow(<SecondComponent />)
+      shallow(<FirstComponent />)
+
+      expectCSSMatches('.sc-a {} .d { color: red; } .sc-b {} .c { color: blue; }')
     })
   })
 })