+++ src/utils/test/generateAlphabeticName.test.js
@@ -1,7 +1,5 @@
 // @flow
-
 import generateAlphabeticName from '../generateAlphabeticName';
-import expect from 'expect';
 
 describe('generateAlphabeticName', () => {
   it('should create alphabetic names for number input data', () => {