+++ src/utils/test/interleave.test.js
@@ -1,5 +1,4 @@
 // @flow
-import expect from 'expect'
 import interleave from '../interleave'
 
 describe('interleave', () => {